feat: added policies

+++ app/Http/Controllers/NoteController.php
@@ -14,6 +14,8 @@ class NoteController extends Controller
 
     public function index()
     {
+        $this->authorize('viewAny', Note::class);
+
         return response()->json([
             'success' => true,
             'data' => NoteResource::collection(Auth::user()->notes)
@@ -22,6 +24,8 @@ class NoteController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create', Note::class);
+
         $note = Note::create([
             'title' => $request->input('title'),
             'content' => $request->input('content'),
@@ -44,6 +48,8 @@ class NoteController extends Controller
 
     public function update(UpdateNoteRequest $request, Note $note)
     {
+        $this->authorize('update', $note);
+
         $note->update($request->validated());
 
         return response()->json([
@@ -54,6 +60,8 @@ class NoteController extends Controller
 
     public function destroy(Note $note)
     {
+        $this->authorize('delete', $note);
+
         $note->delete();
 
         return response()->json([
